Project/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
requests==2.26.0
aiohttp==3.8.1
rich==10.12.0
click==8.0.3
pyyaml==6.0
python-masscan==0.1.1
httpx==0.23.0
pyfiglet==0.8.post1
termcolor==1.1.0
pytest==7.0.1
pytest-asyncio==0.18.1
tenacity==8.2.2
python-nmap==0.7.1
gitleaks-core==8.15.3
wapiti3==3.0.4